pyproject.toml
[tool.poetry]
name = "noah123d"
version = "2025.0.1"
description = "Building assemblies from STL models"
authors = ["Andreas Felix Häberle <felix@42sol.eu>"]
readme = "README.md"
packages = [{include = "noah123d", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^14.1.0"
click = "^8.2.1"
numpy-stl = "^3.2.0"
toml = "^0.10.2"


[tool.poetry.scripts]
noah = "noah123d.__main__:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.7"
behave = "^1.2.6"
pytest-bdd = "^8.1.0"
allure-behave = "^2.15.0"
mkinit = "^1.1.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.0"
mkdocstrings = {extras = ["python"], version = "^0.30.0"}
mkdocstrings-python = "^1.12.0"
mkdocs-autorefs = "^1.2.0"
mkdocs-minify-plugin = "^0.8.0"
markdown-exec = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
